ui/requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.2.0
plotly>=5.20.0
fpdf2>=2.7.8
kaleido>=0.2.1